[ refactor ] enum 클래스 패키지 이동 및 조건 바인딩 클래스 이름 수정 ( #82 )
- `entity` 내에 존재하던 `enum` 클래스를 `enums` 라는 패키지로 이동

- `PaymentCondition` -> `ReqPaymentCondDTO` 로 이름 수정
