+++ src/main/java/com/sparta/gitandrun/payment/contoller/PaymentController.java
@@ -2,7 +2,7 @@ package com.sparta.gitandrun.payment.contoller;
 
 import com.sparta.gitandrun.common.entity.ApiResDto;
 import com.sparta.gitandrun.order.dto.res.ResDto;
-import com.sparta.gitandrun.payment.dto.req.PaymentCondition;
+import com.sparta.gitandrun.payment.dto.req.ReqPaymentCondDTO;
 import com.sparta.gitandrun.payment.dto.req.ReqPaymentPostDTO;
 import com.sparta.gitandrun.payment.dto.res.ResPaymentGetByCustomerDTO;
 import com.sparta.gitandrun.payment.service.PaymentService;
@@ -39,7 +39,7 @@ public class PaymentController {
     @Secured("ROLE_CUSTOMER")
     @GetMapping
     public ResponseEntity<ResDto<ResPaymentGetByCustomerDTO>> readPayment(@AuthenticationPrincipal UserDetailsImpl userDetails,
-                                                                          @RequestBody PaymentCondition condition,
+                                                                          @RequestBody ReqPaymentCondDTO condition,
                                                                           @PageableDefault(sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
 
         return paymentService.getByCustomer(userDetails.getUser(), condition, pageable);
