--- src/main/java/com/sparta/gitandrun/payment/dto/req/PaymentCondition.java
@@ -1,9 +1,0 @@
-package com.sparta.gitandrun.payment.dto.req;
-
-import lombok.Data;
-
-@Data
-public class PaymentCondition {
-    private String paymentStatus;
-    private String sortType;
-}
