+++ src/main/java/com/sparta/gitandrun/payment/entity/enums/SortType.java
@@ -1,4 +1,4 @@
-package com.sparta.gitandrun.payment.entity;
+package com.sparta.gitandrun.payment.entity.enums;
 
 public enum SortType {
     LATEST,
