+++ src/main/java/com/sparta/gitandrun/payment/entity/enums/StatusType.java
@@ -1,4 +1,4 @@
-package com.sparta.gitandrun.payment.entity;
+package com.sparta.gitandrun.payment.entity.enums;
 
 public enum StatusType {
     PAID, CANCEL, ALL;
